+++ changelog.md
@@ -368,4 +368,5 @@ Breaking changes:
 
 ## Next
 
+* Fix: Tests close sequelize connection at end to avoid hang
 * Fix: Typo in changelog